refactor: modularize CLI and improve settings output
