+++ conftest.py
@@ -1,0 +1,9 @@
+"""Ensure project root is on sys.path for tests."""
+from __future__ import annotations
+
+import sys
+from pathlib import Path
+
+ROOT = Path(__file__).resolve().parent
+if str(ROOT) not in sys.path:
+    sys.path.insert(0, str(ROOT))
